Adding program

+++ largest-number/js/new.js
@@ -1,6 +1,6 @@
-var num1 = parseFloat(prompt("Enter first number: "));
-var num2 = parseFloat(prompt("Enter second number: "));
-var num3 = parseFloat(prompt("Enter third number: "));
+var num1 = parseInt(prompt("Enter first number: "));
+var num2 = parseInt(prompt("Enter second number: "));
+var num3 = parseInt(prompt("Enter third number: "));
 var largest;
 
 // check the condition
